ai/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
decorator==5.2.1
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
packaging==26.0
platformdirs==4.4.0
pooch==1.8.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
RapidFuzz==3.13.0
regex==2026.1.15
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
